fixed restartfq in mrt

+++ models/ColorModel.cpp
@@ -318,8 +318,8 @@ void ScaLBL_ColorModel::AssignComponentLabels(double *phase)
         ERROR("Error: ComponentLabels and ComponentAffinity must be the same length! \n");
     }
 
-    int label_count[NLABELS];
-    int label_count_global[NLABELS];
+    double label_count[NLABELS];
+    double label_count_global[NLABELS];
     // Assign the labels
 
     for (int idx=0; idx<NLABELS; idx++) label_count[idx]=0;
@@ -350,7 +350,7 @@ void ScaLBL_ColorModel::AssignComponentLabels(double *phase)
     // Set Dm to match Mask
     //for (int i=0; i<Nx*Ny*Nz; i++) Dm->id[i] = Mask->id[i]; 
 
-    MPI_Allreduce(&label_count[0],&label_count_global[0],NLABELS,MPI_INT,MPI_SUM,Dm->Comm);
+    MPI_Allreduce(&label_count[0],&label_count_global[0],NLABELS,MPI_DOUBLE,MPI_SUM,Dm->Comm);
     //poro=0.0;
 
     if (rank==0) printf("Components labels: %lu \n",NLABELS);
